Title: 최솟값 3개, Time: 122ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 231221/최솟값 3개/top-3-smallest-number.cpp
@@ -17,7 +17,7 @@ int main() {
 
 ll tem_mul;
 
-int run(int n,int tem){
+ll run(int n,int tem){
     int a,b,c;
     if(n<3){
         tem_mul = tem_mul*tem;
